minuitwrp: drm: Await completion of last page flip
If another page flip is scheduled before the last one has been
completed, the DRM API will return EBUSY, causing TWRP to
crash. To avoid this, we can attach an event handler that is called
when the page flip has been completed.

We wait for completion of the page flip directly after scheduling
it. Normally, this is not necessary, however with the current
double-buffering the old buffer may be modified by the renderer
before the flip to the new buffer is completed.

A better solution would be to use triple-buffering, but this is
addressed in a later commit.

Change-Id: Ic4cfe28bdeaa60da76cf121601c68a2077a081cc